Add files via upload
Implemented logic to test the validity of a run. Also uploaded the file being used to quickly test functions and ensure they are working. The next stages are to implement the rest of the scoring logic and to implement a "table" system ( a place for runs and books to be stored).